Improved gossip safety
The cutoff was incorrectly done - if it would cut half-way through a bigger transaction, it would never be able to send the rest of that transaction. This version considers the cutoff as a limit on the number of transactions it can transmit in one msg, just like the package size already is. The user of the library thus controls the size of the msg via the size of the biggest transaction.

ALSO, the TcpTransport can now use a ShieldedDict for ServerIPs.