Removed the doNotSend mechanism
It was not correct - we should avoid sending an item in the reply iff the other side's version is still, at time of composing reply, VectorRelationship.Equal to the current value there. If a new update happens between applying and replying, for example, then doNotSend would falsely exclude that key. Removing this optimization makes the messages a bit less efficient, but is safer regarding correctness.