Removing completed transactions
Initial support for removing items from the gossip backend. It's a pretty strange operation, potentially dangerous. It would probably be better to express it as a quality of the data type, as an interface IDeletable, which the backend would then enforce automatically.

The new removing is used by the consistent backend to remove old transactions. They remain now only in the _freshIndex.