Stronger gossip package limit
The packages can baloon under load, because when a server receives a package, all the items in it get applied within one transactions. To him, all (and only) the interesting stuff in the package becomes one new transaction, although it was maybe 20 to the sender. This causes them to grow, and with them the messages. It actually has a positive effect on tests, but it's better to more strongly limit it.