Removed the Blocked collection
It's really unneccessary. Any incoming message can be applied, even with a transaction currently open and holding that field locked, because a write must have come from some server. Either he accepts it as happening after the commit, and since we are, let's say, in the state Prepared, then we shall pretend so too, or he will fail this transaction. Blocking the incoming msg is thus useless in either case. What matters is that transactions block each other correctly.