Add IDeletable, for clean deleting support
If a TItem implements IDeletable, the GossipBackend will check at every write, and mark it for deletion when it indicates it's deletable. Once deletable it does not take part in the database hash, so that any server can later actually delete it at any time. Every 5 seconds a thread cleans such objects and actually removes them, in a way that each gets kept for at least 5 seconds. Until removal they are readable and get gossiped about normally. The TransactionInfo uses this mechanism, and it's far simpler now.