Eliminating double gossip channels
If two servers open two gossip sessions with each other, one of them (decided using sort order) will skip one reply to kill one of the chains. This eliminates the double gossip chains that the servers used to have, and reduces the overall number of messages with no performance loss.