Add Lww<T>.NextVersion
Creates a guaranteed successor of an Lww. Since server clocks can be out of sync, attempting to write a new version of some Lww could fail just because your clock is behind. This method will in that case use a time that is equal to the old one plus one tick, guaranteeing to win over the old version.