Run and Consistent calls can nest
The biggest issue was that the TransException typically ends up wrapped in an AggregateException if you wait on a nested call. It doesn't matter that the call is actually running on the same thread. Async/await normally takes care of this, but you cannot use it inside a transaction lambda. So, the Shield class now thoroughly digs through TargetInvocation and AggregateExceptions, and performs a rollback and retry if it finds any TransException inside.