Breaking change - Set results reversed
Now the Set methods of both gossip backends return
the relationship of the new value to the old one, which
means that Set returns Greater when your new value has
a greater version than whatever was stored before. Seems
more logical this way, you try to set a certain value and
the result tells you something about that value.