Starting gossip now pushes changes
The opening message can now already contain some
changes, default is 5 latest transactions. This makes the
start msg bigger, but reduces the number of messages
in some cases. It may be turned off by setting the new
config param to 0.

Also, the GossipEnd msg now also has the hash and the
data from the msg it's replying to, so that the receiver of
the GossipEnd may decide to continue the conversation
anyway, in case it has new changes and the hash does
not match any more.