Some optimizations
* The reply transaction now updates _lastSendTime in
same transaction. If two of them conflict, which means
we're trying to send 2 msgs to the same server, then one
will go through, and the one whose transaction retries
will probably not be sent at all. Previous version would
allow both messages through, which is much more
wasteful than a retry.
* Applying the received items is skipped if our DB hash
is equal to the one in the received message.