Revert previous change
Forgot that message transactions do not conflict only with
each other. Since they read the _freshIndex from the end,
they would conflict with any writing transaction. This is
why the _lastSendTime update must be done outside of
any transaction which reads the index. This now includes
the gossip start transaction as well.