Fix IsGossipActive correctness
Since GossipSuppressed direct mail mode depends on
that function, then when we send GossipEnd, we clear
the _lastSendTime in-transaction. The end transaction is
small anyway, and it's now also more reliable - if it rolls
back, the retry will probably send out a GossipReply,
since something probably changed to cause the retry.