Various fixes and improvements
* Restoring ChangedEventArgs.OldValue, because why
not?
* Gossip backends will now throw when given a null
value to save.
* Consistent backend takes care not to barge in on
transactions it was not invited to. This is important to
make sure that once a transaction is resolved it does
not change, and the resolution could change
arbitrarily if new servers suddenly hop on.
* Small fix for state logic of consistent backend.
* GossipBackend takes care to obtain critical info
before calling MergeWith, in case an implementor
makes changes to the "this" object in that method.
Should not, but better to avoid the risk.
* Improved TransactionVector.Merge - it simply
returns the bigger value. Simpler and safer.