Transaction deletable when majority finishes
Analogous to accepting majority vote for commit, we
consider a transaction deletable when a majority has
fully completed it. Waiting for a majority to complete
is safer than marking it deletable as soon as it's done,
and if we wait for all of them, then we might keep old
transactions around for too long.