Make consistent transactions fully consistent
We will check the reads and the writes. A nice
consequence is that the tests no longer need Sleep
calls, since the consistent reading ops are truly
consistent. Even read-only transactions will verify that
they were up-to-date.

Currently, the read fields are locked just like the written
ones. This could be optimized by using an RW-lock
system.