Eliminate busy wait from Consistent_Versioned test
The first run of the transaction locally would sometimes
still not see any value under the key, because the
commit did not yet fully go through on that backend.
Such a transaction run, since it's done in RunConsistent,
would eventually be retried and would see the correct
value. But the Assert inside the transaction would mean
that it would throw on the first run, even though, so to
say, that run would not count. This way is more correct.