Small improvements
* The deletable timer method will now run in two
transactions, to make it less conflicting. Under heavy
load it was struggling to get its job done. The fresh
index is a really strong source of contention...
* The consistent backend state will first release locked
fields and then resolve its tasks. Also, a useless double
call to Complete(false) was removed - rolling back
the continuation already does that as a side-effect.