Use two hash functions for DB hash
Should we ever encounter two different key/version
pairs that hash to the same FNV1a 64bit hash, we could
experience various issues. If one server receives them
both as writes, his DB hash will not change at all due to
those writes. If different servers receive them, they
might think they're in sync, when in fact they're not. This
is very unlikely, and depends on additional factors re.
the exchange of messages between them, but it still
represents a potential issue.

This commit greatly reduces the likelihood of this
happening by changing VersionHash to use two hashes,
one FNV1a 64bit value, and one FNV1a 32bit value.
In the end, this is a 96bit hash that can still collide, so
I'll have to do more work for this. An option is to have
the servers randomly decide to do a full exchange of
all fields sometimes, with a very low likelihood. Or I
might switch to SHA256.