Remove DirectMailRestriction
It was helpful to reduce the number of direct mail msgs
exchanged for one transaction, but it ruined the ability
to compose transactions. This was brought to light
by the other change in this commit - the consistent
backend will not use so many side effects anymore. With
DirectMailRestriction in use, it becomes dangerous to
do anything to two TransactionState fields from one
Shielded transaction - both ops might try changing the
DirectMailRestriction in different ways, with the last one
to try "winning"...