Improve message sizing
Previous version automatically increased gossip message
sizes when under load, but partly because of an accident.
Since it applied all new changes in one transaction, it
would create huge local transactions, and we can only
transmit whole transactions. We now correctly apply
changes as if they were separate transactions on the
receiving server too. So, when the load goes down, they
will send small messages again.

The messages now grow exponentially - when we reply
we try to send twice as much as we received. We will
send more, if under load, because we still try to send all
new changes that occurred since our last msg to this
server in this gossip exchange, which helps keep the
gossip "window" expanding.

Also added a simple feature to ShieldedLocal - that the
GetValueOrDefault method receives a defaultValue
argument.