Fix - ensure that freshness only grows
It was possible for it to go down if the latest items in the
index are deletable, and get removed. Next transaction
would then re-use their old freshness value. This could
lead to updates not being transmitted in gossip, but only
if the deletable timeout was shorter than the gossip
timeout.