Improve consistent transaction handling
Recent change introduced a deadlock which only
occurred rarely. This fixes it. The code is also simpler,
less superstitious. And additionally, a small improvement
is that when a transaction is rejected by another server,
he will send us his rejection and his versions of data in
a single message, and not in two as before.

Also, fixed a minor issue in tests - the port randomization
method would sometimes assign the same port to two
backends.