Non-conflicting time index
Added the class ReverseTimeIndex which encapsulates
the _freshIndex logic, and added a trick to make it less
conflicting - it will perform append as a commute which
just adds the new changes to a singly linked list. The
main transaction is then more likely to complete and
not be serialized with all concurrent transactions.

The queue is hidden from callers, iterating (at least the
descending one) will show items from the queue as if
they were already in the tree. But the adding to the tree
actually happens in a side-effect. This is conflicting, but
the threads help each other there - if two are doing it
in parallel, the one that commits first will have done the
adding for both, and the other will, when repeated, see
there's nothing left to do.

The effect of this is good, but modest. Profiler shows the
CPU usage of the index to be lower, and large messages
get processed more quickly. But seen "from the outside",
the performance is roughly the same as before.

ALSO, fixed a small issue in consistent backend - if a
backend is not participating in a transaction, it ignores it,
but it will now apply it once it reaches the success state.
The difference is minor, perhaps it's even not necessary,
since any transaction success also causes fields to be
updated and transmitted independently too, just because
they're "fresh".