WIP - Fresh index now just a linked list
Implemented like this, the maintenance of the time index
almost completely disappears from the profiler. The 3
node non-consistent test can now process 2x more
transactions in 1 s (now 20k, previous version here 10k),
and the consistent one, if the field number = trans
number, which means if there are no conflicts, runs a bit
under 500 transactions a second. This is probably the best
I can get.

To be able to timely respond to gossip messages, we
now remember where we left of when iterating the time
index, and only reply to messages that are replies to
our previous message to that server.