Several improvements
* Replies will not contain the exact same items we just
received. This was a bit tricky to implement. We determine
which keys from an incoming message completely
overwrote our local values, and were not further changed,
and then we skip them when building a reply.
* ShouldReply reworked. It should be clearer now. Its
decision is now fully based on time info. We remember
our last send time, and the time from the other server's
message that we replied to (his clock). The window
start has matched in all tests so far, so it seems to be
reliable.
* Message growing slightly changed - we do not use
other guy's Items.Length, because in case one server has
no news, and the other has plenty, he would keep sending
10 by 10, the minimal size. Extremely slow. Now every
server sets package size equal to his own last window
size, so the window sizes reliably grow.
* GossipEnd message includes WindowEnd and does
not include LastWindowStart. If they at one point reach
sync, it is assumed that both window starts are 0, meaning
they already have all the older data.

This is all still rather messy, and needs more work.