More strictness to the reply logic
We will now keep state even when we send GossipEnd,
which enables us to accept replies to that End message
correctly, i.e. we will know if they are actually replies to
that message. So, the whole chain is now strictly
controlled, every message we receive we will accept only
if it really is a reply to what we last sent. The only
exceptions are when two servers both send a starter
message to each other, and that weird case when the
other side obviously gave up on the old chain, even
though the gossip state on our side is still OK.

Also, if the message is unexpected, we won't even run
ApplyItems, which saves us time.