Restore strong cut-off
We now use a struct as enumerator for the reverse time
index, which enables cheap cloning and remembering
when a certain freshness value is first encountered. So,
when we would cross the cut-off, we can reset the
enumerator back to the beginning of that freshness block,
and remove the items with that freshness. That way the
cut-off is again strongly respected, and the only way a
message can grow larger if the user runs a transaction
that changes a number of fields that's bigger than it.

This brings this branch to full feature-parity with master,
but I will first write some unit tests for messaging
before merging this there.