Add unit tests + fix a bug
The new test which checks for correct message growth
uncovered a bug. The GossipBackend was using the
window size to determine the size of the next package,
but the window size is a number of transactions, not a
number of key/value pairs. This commit fixes the issue
by adding the package size to the GossipState, which
ensures correct message growth.