New tests, one fails
One of the tests triggered a "chain death", a situation
where the gossip exchange dies and cannot continue. The
issue ultimately comes from complications caused by
allowing GossipEnd messages to be replied to. The servers
end up in incompatible states, and they refuse to answer
to each other. Ultimately, such a chain would time out
and they would be fine, but this is still an issue.

I'll have to consider ways to simplify this logic.