Fixed the gossip end race issue
When sending a GossipStart (which is now a separate
message type, to make things more explicit in the code),
we include any LastTime from the other server exactly for
the case when a GossipEnd gets lost. A GossipStart then
doubles as a GossipEnd - the other server can recognize
that this starter actually starts a new chain and the old one
can safely be dropped.

In the process, I simplified the logic for checking whether
to accept a message. I think it's much more solid now. It
no longer involves any comparisons of time values,
except when checking if our state is obsolete. This means
we are more resilient to any non-monotonous time issues.