Keep expired state, just release enumerator
This may be helpful in situations where delays cause a
gossip chain to timeout on one server, but not yet on the
other. If the first server then sends a GossipStart, he will,
due to the old state being kept, send a corect LastTime
and kill the chain on the other server too.

The only thing that must definitely be released is the
enumerator, because it interferes with the garbage
collector cleaning up old ReverseTimeIndex nodes.