Use ConfigureAwait(false)
The only place where I did not use it is in the Prepare and
RunConsistent methods, since the caller might want their
continuation to run in the same context. It is thus left to
the caller to use ConfigureAwait there, if desired. However,
the internal part of the preparation process still uses it,
since those steps can always run on any thread.