Fix issue with needless reply items check
It did not cover the case when the new message causes no
local changes. Or, it did not do it correctly - it would
be handled by _changeLock not having a value, which I
previously assumed only happens when we were disposed.
But it happens also if there were no changes in the
transaction. In that case, we will simply read the last
freshness in the apply transaction directly.

The fact that the _changeLock does not get set if the
main transaction made no changes is still an issue! Will
write a test for that too.