Eliminate double hashing
The result of calling GetVersionHash on a field value was
immediately passed to another VersionHash.Hash call. So
we now ask IMergeable implementors to simply give us
the bytes of all fields that partake in determining their
version, and we call VersionHash.Hash only once, directly
on all the relevant fields instead of re-hashing a hash.