Use MsgId instead of time + keep TCP connections open
The time in messages was basically just used to match
requests to replies, which is then better done with a
simpler and smaller int ID, since theoretically due to low
clock precision the time might not even change between
two messages.

I originally intended to simply give up on the whole
message matching and delegate this to the ITransport
implementors, since with TCP this whole business is
unnecessary. But this would make it much more difficult
to implement another ITransport. So I only made some
simple changes which enable the TcpTransport to keep
one connection open for an entire gossip exchange.

I might still do this in the future. ITransport implementors
seem like the right place to implement matching of
messages to their replies, since the amount of effort
needed to do it right really depends on the nature of the
transport protocol used.