Add KillGossip message type
When receiving an unexpected GossipReply, the backends
will now respond with a KillGossip message. This helps the
other side drop the invalid state it went into when it send
that reply, otherwise it would stay in it waiting for a reply
we won't send.

Obviously, I decided against delegating the message
chain logic to the transports. The backends depend on the
chain correctness because of the ReverseTimeIndex,
because we remember where we left off iterating in it,
which means we cannot answer any message. So it's not
like the ITransport implementors would have any choice.
It is a strong requirement of the backend itself, and thus
it makes sense that the backend implements it. Done this
way, with the kill message, makes it pretty reliable when
using an unreliable transport, and on the other hand it
does not create overhead when used with a reliable one
like TCP. Checking the messages exchanged in tests that
use the regular TcpTransport reveals that the kill msgs
do not occur.