Improve gossip chain logic
The case when a server sends End and then a new Start.
That server would then no longer expect to receive a
reply to the End message, which is however a possibility.
This change fixes this issue by giving precedence to the
reply to that previous End message.

I think this is done for now. Even this could be called
perfectionism, since eventually their bad gossip states
would time out, resolving the issue. But I wanted to make
the best effort in resolving anything that is reasonably
possible, and I believe the current solution is very strong.
I'd love to do a proper verification, TLA+ model checking
perhaps. But for now, I think the unit tests for gossip
messaging do a good enough job of covering the
different potential failure modes.