Add setting for consistent transaction participants
The consistent backend was using all the transport's
visible servers as transaction participants. In the dodgy
backend asymmetric race test, this was an issue, since the
A server only sees B, but B sees both A and C. By passing
this new argument to the constructor, you can instruct the
consistent backend to use a fixed set of transaction
participants, regardless of which servers are visible to
the transport. The above mentioned test now runs faster.