Use transactional property for consistent participants
This allows you to change the participants while the
backend is running. How exactly this is done and whether
it is safe is not considered right now. Theoretically, one
could use the backend itself as a means of storing the
list of participants and agreeing on changes to it. Or it
could come from outside, from some admin tool. It is
up to the library user to consider the possible
consequences and implement it safely.

Also, added argument validation to GossipBackend.