Fix for consistent initiator voting
The server that initiates transactions always participates
in them, at least because he wants the ability to maybe
fail the transaction after it was prepared. However, if the
transaction participants were explicitly set, and the
initiator is not one of them, then his vote should not count
towards the decision. A new field was added to the
transaction info for this.

Also, a fix for the voting logic - if the number of voters is
even, and we have a tie, then the Reject vote will win.