Enable TransactionParticipants to be transactional
By switching it to an IEnumerable, the users may provide
such an enumerable which gets its value from the
consistent backend itself. In that case, since the thing gets
enumerated in every consistent transaction (except if
empty), it can add additional fields to the transaction.
If the servers are, for example, saved in the backend itself,
then no transaction will succeed unless it's data on
participating servers was up to date.

This is maybe crazy, but the change was really simple.