Add support for removal and expiry
Removal is implemented by setting and transmitting the
Deleted property of a MessageItem. As before, all items
whose Deleted is set do not partake in the DB hash. Expiry
is done via removal. To the library user we hide expired
items, but they are actually non-deleted in the storage
until the deletable clean-up timer finds them, and then
simply performs the usual remove on them. I had to be
very careful, since removal means the DB hash will
change, and servers might not be able to sync if their
values expire at different times. This way, expiry is
irrelevant for the DB hash, only explicit removals affect
it. If an item expires on one server first, they will still
communicate about it as if it were there. If it gets
removed on one server first, then that is something
that gets an entry in the fresh index, gets communicated,
and the other server will then also remove it. Thus, they
are always able to sync.