Decouple deleted from expired state
To enable a very simple use case - if a key expires, you
can now add it back to the backend even if it has not
changed. Previous version might reject the change if the
item has expired, because expired items got deleted,
and a deleted version is considered a higher version.
Now, only if you explicitly delete it we forbid its
reintroduction, at least as long as we still remember its
last state. If it expires, you may reintroduce it.

Another plus is that it's possible to extend the expiry
too, which is also a pretty common use case that the
previous version did not support.

A downside is that the servers can now extend the
expiry if they communicate a lot about the same key
- the ExpireInMs value does not change during
transmission of a message, so the other side can easily
think the expiry was extended when it wasn't. Also, this
is just unpleasantly complicated, and needs a lot more
tests.