Consistent transaction cannot just change expiry
For it to check out, every write it made must be strictly
greater than anything present on other servers. This is
not the case if it just changes expiry, so we won't log
a write in that case. Changing expiry in consistent
transactions will need special handling. For now it is
safer to simply disable it.