Better garbage collection
There are now two intervals, one to specify how often
the clean-up should take place, and another to specify
how long a removable item should linger in the DB
before being finally removed. This allows them to be set
differently, which is also done in this commit - the
default clean-up interval is 2 seconds, and the default
linger time for removables is a full minute, to make it
more likely that a removal is synced with all servers
before actually taking place.