Bugfix + reduce expiry comparison precision
ApplyItems would not compare ExpiresInMs fields when
checking if the two items are completely equal, and it
would skip applying an item whose expiry was just set.

Expiry comparison is now less precise, to reduce the
chance of bogus expiry extensions - if the two values'
difference is less than 100 ms, we will consider them
equal.