Fix removing of expired items
A silly edge case, but could lead to a serious issue - the
item marked Expired has already been excluded from
the DB hash, so we should not perform any change on
the hash if that is the case.

Also changed a sleep length in a clean-up test, because
it would occasionally fail. No issue, just a matter of luck.