Add history match heuristic
When preparing the items for a gossip reply, the backend
now accumulates the effect on the hash of the other
server that these items might have. If that accumulated
value becomes equal to our hash, we stop collecting
older items, and set both window starts to zero. Such a
match can only occur if we already exchanged all old
items and now only have new stuff to gossip about.

This is not exact! For example, if a key had two changes
on our side (in diff transactions), we will only remember
the hash effect of the second one. That's not the effect
that this value will have on the other server - if he
did not see either change, his hash will change by the
effect of both of those changes. Also, in case both had
conflicting changes to a key, our hash effect will
represent the change from our version to the merged
one, but his will represent the change from his to the
merged one. The latter point could, I believe, be
fixed - if we did receive his value, we can calculate
the effect of the merge on his side. But the former issue
cannot be solved without remembering all changes to
the DB, which is impractical.

So, this can only be heuristic, but a useful one. When it
does achieve a match, the backends can save a lot of
effort by not transmitting data which they already
synced before. In the ConsistentRace test the messages
got smaller.