Expiry change now OK in consistent transactions
The fix for consistent transactions means that we can now
send writes which are of equal version, and that means
we can just modify the expiry. Previous version would have
ignored a write that only changes expiry, since that write
would not be Greater on other servers, leading to the
transaction check failing.