Fix for KeyMissing overuse
GetItem was used when iterating over the fresh index, just
to see if a certain item in the index is current. We do not
want to reload any items in that situation, we really just
need a comparison with the current contents in the local
dictionary. Otherwise, clean-up would not work, since
any iteration through the fresh index would refetch the
key. Modified the KeyMissing tests to catch this.