Make TCP connections persistent
Opening and closing them all the time leaves a lot
of connections in the state TIME_WAIT, slowly
consuming OS resources. We now keep a connection
open to any server we know. They now actually have
two connections between them, but it was simpler
this way and it does not require their configs to be
symmetric.

Some issues remain. Changing the ServerIPs dictionary
of the transport is not OK with this version, and can
cause errors.