Completing the persistent connections story
* We now react to changes in the ServerIPs dictionary.
Changed the API to make things simpler - the
constructor receives a dictionary of all servers, and
makes a copy of it in a ShieldedDict. The ShieldedDict
is publicly exposed as a property, and any changes to
it are handled by a PreCommit/SyncSideEffect combo
which updates the persistent connections.
* Added keep-alive. Since they are persistent, we need
this.