Keep-alive safety fix
It was possible for it to run concurrently with another
thread trying to send something, and this is not safe. It
now changes the state to Sending to block anyone from
starting the WriterLoop while it is sending the keep-alive,
and it starts the loop itself if anything gets enqueued
while it was sending.