Some more ConfigureAwaits
And a comment, since I almost forgot why I had to put the
Connect call in a Task.Run - trying to run any await on an
old ASP.NET WebForms thread causes an exception. So even
though we could start the initial Connect call on the same
thread, we are forced to pass it to the pool to avoid this
problem.

BTW, the new ConfigureAwaits are not actually needed, since
the TcpClientConn runs everything on pool threads, but just in
case the code will ever be changed, so that I don't have to
think about it...