Some more ConfigureAwaits
I thought if the Prepare and RunConsistent methods do not
use ConfigureAwait, they will "inherit" the choice from their
caller. But of course, that's not how that works. If you
execute RunConsistent(...).ConfigureAwait(false), the awaits
inside Prepare will still see the context and capture it, and
any later retries of your lambda would be running in the
captured context. So I have to accept
runTransOnCapturedContext as an argument, if I want to offer people the choice. The default is to capture, so that your
transaction lambda and any side-effects run in the captured
context (e.g. on the UI thread). This matches the .NET default.

ConfigureAwait is really the worst!! Though, I must admit,
even if it worked in the opposite way, by default not
capturing the context, in my case here I would still need this
new argument...