Fix a tricky bug in consistent mode
The new stress test runs transactions over two fields each,
and uses DirectMail.Always just to make things worse. It
revealed a bug, where sometimes it would detect that a
field received conflicting writes. The bug was in
ApplyAndSetSuccess, which would not apply the changes of
a successful transaction if it could not read the info from
the wrapped backend. This is a mistake, since the info
becomes deletable at some point. The server would then
skip applying changes from the transaction, but would
release the field locks that the transaction held.

Also changed:
* ApplyMethods uses Expressions to be easier to use.
* Slight change in how transaction check treats deletes.
* Refactored the TransactionalTickCount out of
MessageItem.
* Transactions now deletable as soon as they are done.
Improves performance quite a bit, saves updating the
state of decided transactions. This info is not used
currently anyway.
* Better handling of transactions that were forgotten by
their initiator - if he succeeded previously, he must make
sure to bring others up to date when telling them to fail
now.