Add byte limit for messages
In case some values will be very large, the gossip backend
now also limits messages by bytes. It only counts the number
of bytes it sees - the total bytes of serialized values. The actual
message will be larger, due to the metadata sent along with
that. By default the limit is 10 MB, but configurable, of course.