Suppress execution context flow to TcpClientConn tasks
The amount of footguns in the .NET async programming
model is astounding. I noticed, when checking the log output,
that various message logging was happening seemingly in
the scope of some random transaction. In fact, the logging
scope, being AsyncLocal, "flowed over" to the tasks that were
started by the TcpClientConnection when it began connecting,
and since then all log messages related to the connection
would log in that scope.

The whole async story is just so simple and magical, but in the
end you cannot do anything with it safely without knowing
exactly what magic is happening where. That's bad magic.

Also, the TaskCompletionSource has a footgun as well -
TrySetResult runs continuations synchronously by default.
I already noticed this some time ago, when I had to put
all TrySetResult calls in side-effects, which is after all logical.
But I thought it's only sometimes... Well, I'm using the option
to force it not to do that. Most of the time it is actually fine,
when we signal e.g. transaction prepared, we want the logic
to run ASAP and decide whether to commit or whatnot. But
in case the continuation decides to be slow, it would slow
down the GossipBackend reply processes, so it seems wiser
to avoid this default as well.

And yes, added a console test project, just to see the log...