Various fixes and improvements
* VectorBase now demands inheritors to implement GetBytes,
Compare and Merge. This corresponds to the IMergeable
interface. It is simpler than demanding the value be an
IMergeable implementor. This was done so that we no longer
use default .NET comparers and the default .NET hash code
calculation! It might not give the same result on different
servers.
* BitConverter is also not used directly anymore for the same
reason. SafeBitConverter will always return bytes in
little-endian order.
* ReverseTimeIndex AppendItem simplified. This is much safer,
and does not impact performance.

I am still struggling to figure out why the test in the console
app occasionally logs that a gossip exchange failed. It is not
a critical issue, the next exchange between the same servers
succeeds, and the test completes successfully. But, I believe
this should not be possible, so it worries me a bit.