First attempt at a Paxos-like consistent algorithm
Awfully slow, and also very buggy. This version here seems to
be safe - the tests are not detecting any inconsistencies in the
data store - but it is showing weird issues with the internal
state.

The algorithm is inspired by CASPaxos, but it is a mess. I am not
really sure why the rules in CanMakeProgress work this way, and
it is probably wrong. But the basic idea is to have them vote like
in Paxos, while each of them plays the role of proposer and
acceptor. Instead of having new rounds for resolving some types
of conflicts, the servers simply switch their vote, but in a way
that matches how a proposer in a new round would pick a new
value to propose...