Improve Paxos-style consistent transactions
This is now relatively performant, and I am more confident that
it is safe. The changes here mostly improve its performance.

This is a kind of poor man's Paxos. It is the most trivial voting
scheme, with only one hack - a server in Accepted state for one
transaction will switch to another if it has a higher ballot number
and already gained an Accepted vote from any other server. This
will resolve some race conditions correctly (we know the old
transaction can never succeed because the new one could have
only gained that Accepted vote if it had a majority of Promised
votes...).

I can think of situations where this algorithm does get blocked
on the majority side of a partition. Full-blown Paxos would
launch a new round and resolve the block, but this one has no
timeouts and no concept of new transactions helping old ones.
However, it is pretty simple, and better than the 2PC version.