Add support for cancelling retries of consistent transactions
The RunConsistent method receives an optional cancellation
token, and if it gets cancelled, it will only continue until the
current attempt finishes. This should be the main way to control
the duration now, so I increased default attempts to 100.

If a user wants to stop waiting immediately, that is easy to
implement with existing .NET tools. I did not want to do
such immediate cancellation here, since I think it's better that
the API reflects the reality of how it works - you cannot
reliably cancel an attempt that is in progress. Other servers may
decide to commit while your server is disconnected from them.