Introduce ConsistentOutcome + switch to hard cancellation
An enum that returns the actual outcome of a consistent
transaction. I guess a lot of people will prefer to avoid
exceptions, so this leaves the choice to lib users. The overload
that returns a value now returns a struct which wraps the value
and the outcome.

Cancellation is now immediate - if an attempt was in progress,
we will immediately stop waiting for it. The outcome will
indicate whether that was the case. This change is important
since transactions can possibly get stuck forever due to
network issues, so this is probably way more practical. Also
it's easy to set a timeout on transactions now, using the
CancellationTokenSource.