Remove BallotNumber
I realized that this is really not like Paxos :) And the ballot
numbers are just useless overhead. The transactions are now
prioritized based on their ID sort order, which is extremely
simple and fairer than using initiator ID as I did in the 2PC
variant.

Also removed the Apply call for transactions where we are
not in the participating servers. The server will just receive the
committed writes eventually via Gossip.